Nouvel obstacle Boite, création Source dans evolue(), correction classe Dalle, raylib P11-v2

+++ commun/Aleatoire/Aleatoire.h
@@ -7,7 +7,7 @@ class Aleatoire
 public:
   // constructeurs
   Aleatoire(int graine)
-  : generateur(graine)
+  : generateur(static_cast<unsigned int>(graine))
   {}
 
   Aleatoire()
